Kill intermediate types before rendering template
There were intermediate data stuctures being used to resolve models to
strings before handing them off to templates in case they triggered an
error because of missing reference models.

Instead of an error we return serde_json::Value. It'd be nice to have
this error but it's making the code needlessly complex. Maybe if askama
supports errors we can bring the error back.